Externalize build methods from EntityPlacerConfigs
