+++ optaplanner-core/src/main/java/org/optaplanner/core/config/constructionheuristic/placer/QueuedEntityPlacerConfig.java
@@ -17,15 +17,11 @@
 package org.optaplanner.core.config.constructionheuristic.placer;
 
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlElements;
 
-import org.optaplanner.core.config.heuristic.selector.common.SelectionCacheType;
-import org.optaplanner.core.config.heuristic.selector.common.SelectionOrder;
 import org.optaplanner.core.config.heuristic.selector.entity.EntitySelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.MoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.composite.CartesianProductMoveSelectorConfig;
@@ -40,22 +36,16 @@ import org.optaplanner.core.config.heuristic.selector.move.generic.chained.KOptM
 import org.optaplanner.core.config.heuristic.selector.move.generic.chained.SubChainChangeMoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.generic.chained.SubChainSwapMoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.generic.chained.TailChainSwapMoveSelectorConfig;
-import org.optaplanner.core.config.heuristic.selector.value.ValueSelectorConfig;
 import org.optaplanner.core.config.util.ConfigUtils;
-import org.optaplanner.core.impl.constructionheuristic.placer.QueuedEntityPlacer;
-import org.optaplanner.core.impl.domain.entity.descriptor.EntityDescriptor;
-import org.optaplanner.core.impl.domain.variable.descriptor.GenuineVariableDescriptor;
+import org.optaplanner.core.impl.constructionheuristic.placer.QueuedEntityPlacerFactory;
 import org.optaplanner.core.impl.heuristic.HeuristicConfigPolicy;
-import org.optaplanner.core.impl.heuristic.selector.entity.EntitySelector;
-import org.optaplanner.core.impl.heuristic.selector.move.MoveSelector;
-import org.optaplanner.core.impl.heuristic.selector.move.MoveSelectorFactory;
 
 public class QueuedEntityPlacerConfig extends EntityPlacerConfig<QueuedEntityPlacerConfig> {
 
     public static QueuedEntityPlacerConfig unfoldNew(HeuristicConfigPolicy configPolicy,
             List<MoveSelectorConfig> templateMoveSelectorConfigList) {
         QueuedEntityPlacerConfig config = new QueuedEntityPlacerConfig();
-        config.entitySelectorConfig = config.buildEntitySelectorConfig(configPolicy);
+        config.entitySelectorConfig = new QueuedEntityPlacerFactory(config).buildEntitySelectorConfig(configPolicy);
         config.moveSelectorConfigList = new ArrayList<>(templateMoveSelectorConfigList.size());
         List<MoveSelectorConfig> leafMoveSelectorConfigList = new ArrayList<>(templateMoveSelectorConfigList.size());
         for (MoveSelectorConfig templateMoveSelectorConfig : templateMoveSelectorConfigList) {
@@ -108,9 +98,6 @@ public class QueuedEntityPlacerConfig extends EntityPlacerConfig<QueuedEntityPla
     })
     protected List<MoveSelectorConfig> moveSelectorConfigList = null;
 
-    public QueuedEntityPlacerConfig() {
-    }
-
     public EntitySelectorConfig getEntitySelectorConfig() {
         return entitySelectorConfig;
     }
@@ -130,88 +117,6 @@ public class QueuedEntityPlacerConfig extends EntityPlacerConfig<QueuedEntityPla
     // ************************************************************************
     // Builder methods
     // ************************************************************************
-
-    @Override
-    public QueuedEntityPlacer buildEntityPlacer(HeuristicConfigPolicy configPolicy) {
-        EntitySelectorConfig entitySelectorConfig_ = buildEntitySelectorConfig(configPolicy);
-        EntitySelector entitySelector = entitySelectorConfig_.buildEntitySelector(configPolicy,
-                SelectionCacheType.PHASE, SelectionOrder.ORIGINAL);
-
-        List<MoveSelectorConfig> moveSelectorConfigList_;
-        if (ConfigUtils.isEmptyCollection(moveSelectorConfigList)) {
-            EntityDescriptor entityDescriptor = entitySelector.getEntityDescriptor();
-            Collection<GenuineVariableDescriptor> variableDescriptors = entityDescriptor.getGenuineVariableDescriptors();
-            List<MoveSelectorConfig> subMoveSelectorConfigList = new ArrayList<>(
-                    variableDescriptors.size());
-            for (GenuineVariableDescriptor variableDescriptor : variableDescriptors) {
-                subMoveSelectorConfigList.add(buildChangeMoveSelectorConfig(
-                        configPolicy, entitySelectorConfig_.getId(), variableDescriptor));
-            }
-            MoveSelectorConfig subMoveSelectorConfig;
-            if (subMoveSelectorConfigList.size() > 1) {
-                // Default to cartesian product (not a queue) of planning variables.
-                subMoveSelectorConfig = new CartesianProductMoveSelectorConfig(subMoveSelectorConfigList);
-            } else {
-                subMoveSelectorConfig = subMoveSelectorConfigList.get(0);
-            }
-            moveSelectorConfigList_ = Collections.singletonList(subMoveSelectorConfig);
-        } else {
-            moveSelectorConfigList_ = moveSelectorConfigList;
-        }
-        List<MoveSelector> moveSelectorList = new ArrayList<>(moveSelectorConfigList_.size());
-        for (MoveSelectorConfig moveSelectorConfig : moveSelectorConfigList_) {
-            MoveSelector moveSelector = MoveSelectorFactory.create(moveSelectorConfig)
-                    .buildMoveSelector(configPolicy, SelectionCacheType.JUST_IN_TIME, SelectionOrder.ORIGINAL);
-            moveSelectorList.add(moveSelector);
-        }
-        return new QueuedEntityPlacer(entitySelector, moveSelectorList);
-    }
-
-    private EntitySelectorConfig buildEntitySelectorConfig(HeuristicConfigPolicy configPolicy) {
-        EntitySelectorConfig entitySelectorConfig_;
-        if (entitySelectorConfig == null) {
-            entitySelectorConfig_ = new EntitySelectorConfig();
-            EntityDescriptor entityDescriptor = configPolicy.getSolutionDescriptor().deduceEntityDescriptor(null);
-            Class<?> entityClass = entityDescriptor.getEntityClass();
-            entitySelectorConfig_.setId(entityClass.getName());
-            entitySelectorConfig_.setEntityClass(entityClass);
-            if (EntitySelectorConfig.hasSorter(configPolicy.getEntitySorterManner(), entityDescriptor)) {
-                entitySelectorConfig_.setCacheType(SelectionCacheType.PHASE);
-                entitySelectorConfig_.setSelectionOrder(SelectionOrder.SORTED);
-                entitySelectorConfig_.setSorterManner(configPolicy.getEntitySorterManner());
-            }
-        } else {
-            entitySelectorConfig_ = entitySelectorConfig;
-        }
-        if (entitySelectorConfig_.getCacheType() != null
-                && entitySelectorConfig_.getCacheType().compareTo(SelectionCacheType.PHASE) < 0) {
-            throw new IllegalArgumentException("The queuedEntityPlacer (" + this
-                    + ") cannot have an entitySelectorConfig (" + entitySelectorConfig_
-                    + ") with a cacheType (" + entitySelectorConfig_.getCacheType()
-                    + ") lower than " + SelectionCacheType.PHASE + ".");
-        }
-        return entitySelectorConfig_;
-    }
-
-    private ChangeMoveSelectorConfig buildChangeMoveSelectorConfig(HeuristicConfigPolicy configPolicy,
-            String entitySelectorConfigId, GenuineVariableDescriptor variableDescriptor) {
-        ChangeMoveSelectorConfig changeMoveSelectorConfig = new ChangeMoveSelectorConfig();
-        changeMoveSelectorConfig.setEntitySelectorConfig(
-                EntitySelectorConfig.newMimicSelectorConfig(entitySelectorConfigId));
-        ValueSelectorConfig changeValueSelectorConfig = new ValueSelectorConfig();
-        changeValueSelectorConfig.setVariableName(variableDescriptor.getVariableName());
-        if (ValueSelectorConfig.hasSorter(configPolicy.getValueSorterManner(), variableDescriptor)) {
-            if (variableDescriptor.isValueRangeEntityIndependent()) {
-                changeValueSelectorConfig.setCacheType(SelectionCacheType.PHASE);
-            } else {
-                changeValueSelectorConfig.setCacheType(SelectionCacheType.STEP);
-            }
-            changeValueSelectorConfig.setSelectionOrder(SelectionOrder.SORTED);
-            changeValueSelectorConfig.setSorterManner(configPolicy.getValueSorterManner());
-        }
-        changeMoveSelectorConfig.setValueSelectorConfig(changeValueSelectorConfig);
-        return changeMoveSelectorConfig;
-    }
 
     @Override
     public QueuedEntityPlacerConfig inherit(QueuedEntityPlacerConfig inheritedConfig) {
