+++ optaplanner-core/src/main/java/org/optaplanner/core/config/constructionheuristic/placer/QueuedValuePlacerConfig.java
@@ -19,10 +19,6 @@ package org.optaplanner.core.config.constructionheuristic.placer;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlElements;
 
-import org.optaplanner.core.api.domain.valuerange.ValueRangeProvider;
-import org.optaplanner.core.config.heuristic.selector.common.SelectionCacheType;
-import org.optaplanner.core.config.heuristic.selector.common.SelectionOrder;
-import org.optaplanner.core.config.heuristic.selector.entity.EntitySelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.MoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.composite.CartesianProductMoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.composite.UnionMoveSelectorConfig;
@@ -38,19 +34,10 @@ import org.optaplanner.core.config.heuristic.selector.move.generic.chained.SubCh
 import org.optaplanner.core.config.heuristic.selector.move.generic.chained.TailChainSwapMoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.value.ValueSelectorConfig;
 import org.optaplanner.core.config.util.ConfigUtils;
-import org.optaplanner.core.impl.constructionheuristic.placer.QueuedValuePlacer;
-import org.optaplanner.core.impl.domain.entity.descriptor.EntityDescriptor;
-import org.optaplanner.core.impl.domain.variable.descriptor.GenuineVariableDescriptor;
-import org.optaplanner.core.impl.heuristic.HeuristicConfigPolicy;
-import org.optaplanner.core.impl.heuristic.selector.move.MoveSelector;
-import org.optaplanner.core.impl.heuristic.selector.move.MoveSelectorFactory;
-import org.optaplanner.core.impl.heuristic.selector.value.EntityIndependentValueSelector;
-import org.optaplanner.core.impl.heuristic.selector.value.ValueSelector;
 
 public class QueuedValuePlacerConfig extends EntityPlacerConfig<QueuedValuePlacerConfig> {
 
-    public static QueuedValuePlacerConfig unfoldNew(HeuristicConfigPolicy configPolicy,
-            MoveSelectorConfig templateMoveSelectorConfig) {
+    public static QueuedValuePlacerConfig unfoldNew(MoveSelectorConfig templateMoveSelectorConfig) {
         throw new UnsupportedOperationException("The <constructionHeuristic> contains a moveSelector ("
                 + templateMoveSelectorConfig + ") and the <queuedValuePlacer> does not support unfolding those yet.");
     }
@@ -81,9 +68,6 @@ public class QueuedValuePlacerConfig extends EntityPlacerConfig<QueuedValuePlace
     })
     private MoveSelectorConfig moveSelectorConfig = null;
 
-    public QueuedValuePlacerConfig() {
-    }
-
     public Class<?> getEntityClass() {
         return entityClass;
     }
@@ -106,76 +90,6 @@ public class QueuedValuePlacerConfig extends EntityPlacerConfig<QueuedValuePlace
 
     public void setMoveSelectorConfig(MoveSelectorConfig moveSelectorConfig) {
         this.moveSelectorConfig = moveSelectorConfig;
-    }
-
-    // ************************************************************************
-    // Builder methods
-    // ************************************************************************
-
-    @Override
-    public QueuedValuePlacer buildEntityPlacer(HeuristicConfigPolicy configPolicy) {
-        EntityDescriptor entityDescriptor = configPolicy.getSolutionDescriptor().deduceEntityDescriptor(entityClass);
-        ValueSelectorConfig valueSelectorConfig_ = buildValueSelectorConfig(configPolicy, entityDescriptor);
-        ValueSelector valueSelector = valueSelectorConfig_.buildValueSelector(configPolicy, entityDescriptor,
-                SelectionCacheType.PHASE, SelectionOrder.ORIGINAL, false);
-
-        MoveSelectorConfig moveSelectorConfig_ = moveSelectorConfig == null ? buildChangeMoveSelectorConfig(configPolicy,
-                valueSelectorConfig_.getId(), valueSelector.getVariableDescriptor()) : moveSelectorConfig;
-
-        MoveSelector moveSelector = MoveSelectorFactory.create(moveSelectorConfig_)
-                .buildMoveSelector(configPolicy, SelectionCacheType.JUST_IN_TIME, SelectionOrder.ORIGINAL);
-        if (!(valueSelector instanceof EntityIndependentValueSelector)) {
-            throw new IllegalArgumentException("The queuedValuePlacer (" + this
-                    + ") needs to be based on an EntityIndependentValueSelector (" + valueSelector + ")."
-                    + " Check your @" + ValueRangeProvider.class.getSimpleName() + " annotations.");
-
-        }
-        return new QueuedValuePlacer((EntityIndependentValueSelector) valueSelector, moveSelector);
-    }
-
-    private ValueSelectorConfig buildValueSelectorConfig(HeuristicConfigPolicy configPolicy,
-            EntityDescriptor entityDescriptor) {
-        ValueSelectorConfig valueSelectorConfig_;
-        if (valueSelectorConfig == null) {
-            valueSelectorConfig_ = new ValueSelectorConfig();
-            Class<?> entityClass = entityDescriptor.getEntityClass();
-            GenuineVariableDescriptor variableDescriptor = entityDescriptor.deduceVariableDescriptor(null);
-            valueSelectorConfig_.setId(entityClass.getName() + "." + variableDescriptor.getVariableName());
-            valueSelectorConfig_.setVariableName(variableDescriptor.getVariableName());
-            if (ValueSelectorConfig.hasSorter(configPolicy.getValueSorterManner(), variableDescriptor)) {
-                valueSelectorConfig_.setCacheType(SelectionCacheType.PHASE);
-                valueSelectorConfig_.setSelectionOrder(SelectionOrder.SORTED);
-                valueSelectorConfig_.setSorterManner(configPolicy.getValueSorterManner());
-            }
-        } else {
-            valueSelectorConfig_ = valueSelectorConfig;
-        }
-        if (valueSelectorConfig_.getCacheType() != null
-                && valueSelectorConfig_.getCacheType().compareTo(SelectionCacheType.PHASE) < 0) {
-            throw new IllegalArgumentException("The queuedValuePlacer (" + this
-                    + ") cannot have a valueSelectorConfig (" + valueSelectorConfig_
-                    + ") with a cacheType (" + valueSelectorConfig_.getCacheType()
-                    + ") lower than " + SelectionCacheType.PHASE + ".");
-        }
-        return valueSelectorConfig_;
-    }
-
-    private ChangeMoveSelectorConfig buildChangeMoveSelectorConfig(HeuristicConfigPolicy configPolicy,
-            String valueSelectorConfigId, GenuineVariableDescriptor variableDescriptor) {
-        ChangeMoveSelectorConfig changeMoveSelectorConfig = new ChangeMoveSelectorConfig();
-        EntitySelectorConfig changeEntitySelectorConfig = new EntitySelectorConfig();
-        EntityDescriptor entityDescriptor = variableDescriptor.getEntityDescriptor();
-        changeEntitySelectorConfig.setEntityClass(entityDescriptor.getEntityClass());
-        if (EntitySelectorConfig.hasSorter(configPolicy.getEntitySorterManner(), entityDescriptor)) {
-            changeEntitySelectorConfig.setCacheType(SelectionCacheType.PHASE);
-            changeEntitySelectorConfig.setSelectionOrder(SelectionOrder.SORTED);
-            changeEntitySelectorConfig.setSorterManner(configPolicy.getEntitySorterManner());
-        }
-        changeMoveSelectorConfig.setEntitySelectorConfig(changeEntitySelectorConfig);
-        ValueSelectorConfig changeValueSelectorConfig = new ValueSelectorConfig();
-        changeValueSelectorConfig.setMimicSelectorRef(valueSelectorConfigId);
-        changeMoveSelectorConfig.setValueSelectorConfig(changeValueSelectorConfig);
-        return changeMoveSelectorConfig;
     }
 
     @Override
