+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/constructionheuristic/placer/AbstractEntityPlacerFactory.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.optaplanner.core.impl.constructionheuristic.placer;
+
+import org.optaplanner.core.config.constructionheuristic.placer.EntityPlacerConfig;
+import org.optaplanner.core.config.heuristic.selector.common.SelectionCacheType;
+import org.optaplanner.core.config.heuristic.selector.common.SelectionOrder;
+import org.optaplanner.core.config.heuristic.selector.entity.EntitySelectorConfig;
+import org.optaplanner.core.config.heuristic.selector.move.generic.ChangeMoveSelectorConfig;
+import org.optaplanner.core.config.heuristic.selector.value.ValueSelectorConfig;
+import org.optaplanner.core.impl.domain.variable.descriptor.GenuineVariableDescriptor;
+import org.optaplanner.core.impl.heuristic.HeuristicConfigPolicy;
+
+abstract class AbstractEntityPlacerFactory<Config_ extends EntityPlacerConfig> implements EntityPlacerFactory {
+
+    protected final Config_ placerConfig;
+
+    protected AbstractEntityPlacerFactory(Config_ placerConfig) {
+        this.placerConfig = placerConfig;
+    }
+
+    protected ChangeMoveSelectorConfig buildChangeMoveSelectorConfig(HeuristicConfigPolicy configPolicy,
+            String entitySelectorConfigId, GenuineVariableDescriptor variableDescriptor) {
+        ChangeMoveSelectorConfig changeMoveSelectorConfig = new ChangeMoveSelectorConfig();
+        changeMoveSelectorConfig.setEntitySelectorConfig(
+                EntitySelectorConfig.newMimicSelectorConfig(entitySelectorConfigId));
+        ValueSelectorConfig changeValueSelectorConfig = new ValueSelectorConfig();
+        changeValueSelectorConfig.setVariableName(variableDescriptor.getVariableName());
+        if (ValueSelectorConfig.hasSorter(configPolicy.getValueSorterManner(), variableDescriptor)) {
+            if (variableDescriptor.isValueRangeEntityIndependent()) {
+                changeValueSelectorConfig.setCacheType(SelectionCacheType.PHASE);
+            } else {
+                changeValueSelectorConfig.setCacheType(SelectionCacheType.STEP);
+            }
+            changeValueSelectorConfig.setSelectionOrder(SelectionOrder.SORTED);
+            changeValueSelectorConfig.setSorterManner(configPolicy.getValueSorterManner());
+        }
+        changeMoveSelectorConfig.setValueSelectorConfig(changeValueSelectorConfig);
+        return changeMoveSelectorConfig;
+    }
+}
